Features: implements payList
implemento la lista de pagos, y creo un trigger en la bd, que cuando elimino un pago se le suma al cliente lo que debe, como si no hubiese realizado el pago

+++ app-react/src/Components/ClientForm/Compra.jsx
@@ -40,8 +40,8 @@ export const Compra = ({compras, client, onInsertBuy}) => {
     const handleSubmit = (e) => {
         e.preventDefault();
         //AGREGAMOS SUBTOTAL Y TOTAL
-        buy.total=(parseInt(buy.precio)*parseInt(buy.cantidad));
-        buy.subtotal=(parseInt(buy.precio));
+        buy.total=(parseFloat(buy.precio)*parseFloat(buy.cantidad));
+        buy.subtotal=(parseFloat(buy.precio));
         console.log("ENVIAMOS: ",buy);
         //INSERTAMOS EL TOTAL
         onInsertBuy(buy);
